+++ kernel/src/abi/linux/riscv64/mod.rs
@@ -78,7 +78,7 @@ impl AbiModule for LinuxRiscv64Abi {
     fn handle_event(
         &mut self,
         event: crate::ipc::Event,
-        target_task_id: u32,
+        target_task_id: usize,
     ) -> Result<EventProcessOutcome, &'static str> {
         generic::signal::handle_event_for_task(
             &self.0,
